Renamed src/physics/common -> src/physics/share, updated refs.
