+++ components/scream/src/physics/CMakeLists.txt
@@ -1,4 +1,4 @@
-add_subdirectory(common)
+add_subdirectory(share)
 add_subdirectory(p3)
 add_subdirectory(rrtmgp)
 add_subdirectory(shoc)
