+++ components/scream/src/physics/p3/CMakeLists.txt
@@ -63,11 +63,11 @@ endif()
 # link_directories(${SCREAM_TPL_LIBRARY_DIRS} ${SCREAM_LIBRARY_DIRS})
 add_library(p3 ${P3_SRCS})
 target_include_directories(p3 PUBLIC ${SCREAM_INCLUDE_DIRS})
-target_include_directories(p3 PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/../common)
+target_include_directories(p3 PUBLIC ${CMAKE_CURRENT_SOURCE_DIR}/../share)
 target_include_directories(p3 SYSTEM PUBLIC ${SCREAM_TPL_INCLUDE_DIRS} ${CIMEROOT}/src/share/include)
 set_target_properties(p3 PROPERTIES
   Fortran_MODULE_DIRECTORY ${SCREAM_F90_MODULES})
-target_link_libraries(p3 physics_common scream_share ${SCREAM_TPL_LIBRARIES})
+target_link_libraries(p3 physics_share scream_share ${SCREAM_TPL_LIBRARIES})
 
 if (NOT SCREAM_LIB_ONLY)
   add_subdirectory(tests)
