+++ components/scream/src/physics/p3/tests/p3_saturation_unit_tests.cpp
@@ -7,8 +7,8 @@
 #include "ekat/util/scream_kokkos_utils.hpp"
 #include "physics/p3/p3_functions.hpp"
 #include "physics/p3/p3_functions_f90.hpp"
-#include "physics/common/physics_functions.hpp"
-#include "physics/common/physics_saturation_impl.hpp"
+#include "physics/share/physics_functions.hpp"
+#include "physics/share/physics_saturation_impl.hpp"
 
 #include "p3_unit_tests_common.hpp"
